Move leaderboard to public api
