+++ src/CS2/Leaderboard/LeaderboardCollection.cs
@@ -1,4 +1,6 @@
 using GangsAPI.Extensions;
+using GangsAPI.Services;
+using GangsAPI.Services.Server;
 using Microsoft.Extensions.DependencyInjection;
 
 namespace Leaderboard;
