+++ src/GangsAPI/Services/ILeaderboard.cs
@@ -1,6 +1,4 @@
-using GangsAPI;
-
-namespace Leaderboard;
+namespace GangsAPI.Services;
 
 public interface ILeaderboard : IPluginBehavior {
   Task<IEnumerable<(int, double)>> GetTopGangs(int limit = 10, int offset = 0);
